add new fetures and relashionship and model

+++ database/migrations/2022_05_28_122041_create_products_table.php
@@ -18,14 +18,14 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('image');
-            $table->integer('price');
+            $table->float('price');
             $table->float('stock');
             $table->timestamps();
 
         });
     }
 
-    
+
 
     /**
      * Reverse the migrations.
